https://github.com/javers/javers/issues/666 refactor package name org.javers.spring.auditable.aspect.springdata -> org.javers.spring.auditable.aspect.springdatajpa.
